requirements.txt
asgiref==3.4.1
autopep8==1.5.7
coverage==5.5
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-taggit==1.5.1
flake8==3.9.2
mccabe==0.6.1
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
